requirements.txt
monai
jupyterlab
pillow
matplotlib
glob2
scikit-learn
torchvision
pynvml
transformers
segmentation_models_pytorch
albumentations
pydantic